requirements.txt
torch
librosa
pandas
numpy
pyyaml
soundfile
scikit-learn
matplotlib
fastapi
uvicorn
pytest
wandb
optuna
